competed week 2 error free
I came back and fixed the errors on programmer humour and restarted dog photo with a clearer mind and completed it without issues this time!
did not understand the need for. second button as per instructions, but I created a use for it.
NB- need to learn error handling better